Save planner

+++ Backend/pi2/src/main/java/com/example/pi2/service/PlannerService.java
@@ -2,6 +2,7 @@ package com.example.pi2.service;
 
 import com.example.pi2.domain.PlannerDTO;
 import com.example.pi2.domain.PlannerDtoToCsv;
+import com.example.pi2.domain.PlannerWithRecipeDTO;
 import com.example.pi2.model.Planner;
 import com.example.pi2.model.Recipe;
 import com.example.pi2.repository.PlannerRepository;
@@ -38,6 +39,12 @@ public class PlannerService {
 	  public List<Planner> findByUserId(String idUser) {
 
 			return plannerRepository.findByUserId(idUser);
+
+	  }
+
+	  public PlannerWithRecipeDTO findPlanByUserId(String idUser) {
+
+			return mapperPlan(plannerRepository.findOneByUser(idUser));
 
 	  }
 
@@ -92,40 +99,43 @@ public class PlannerService {
 //			PlannerDtoToCsv plannerDTO = new PlannerDtoToCsv();
 			List<PlannerDtoToCsv> plannerDtoToCsvList = new ArrayList<>();
 
-			if (planner.getSunday() != null) {
-				  PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getSunday());
-				  plannerDtoToCsv.setDayOfTheWeek("Domingo");
-				  plannerDtoToCsvList.add(plannerDtoToCsv);
-			}
-			if (planner.getMonday() != null) {
-				  PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getMonday());
-				  plannerDtoToCsv.setDayOfTheWeek("Lunes");
-				  plannerDtoToCsvList.add(plannerDtoToCsv);
-			}
-			if (planner.getTuesday() != null) {
-				  PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getTuesday());
-				  plannerDtoToCsv.setDayOfTheWeek("Martes");
-				  plannerDtoToCsvList.add(plannerDtoToCsv);
-			}
-			if (planner.getWednesday() != null) {
-				  PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getWednesday());
-				  plannerDtoToCsv.setDayOfTheWeek("Miércoles");
-				  plannerDtoToCsvList.add(plannerDtoToCsv);
-			}
-			if (planner.getThursday() != null) {
-				  PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getThursday());
-				  plannerDtoToCsv.setDayOfTheWeek("Jueves");
-				  plannerDtoToCsvList.add(plannerDtoToCsv);
-			}
-			if (planner.getFriday() != null) {
-				  PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getFriday());
-				  plannerDtoToCsv.setDayOfTheWeek("Viernes");
-				  plannerDtoToCsvList.add(plannerDtoToCsv);
-			}
-			if (planner.getSaturday() != null) {
-				  PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getSaturday());
-				  plannerDtoToCsv.setDayOfTheWeek("Sábado");
-				  plannerDtoToCsvList.add(plannerDtoToCsv);
+			if (planner != null) {
+
+				  if (planner.getSunday() != null) {
+						PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getSunday());
+						plannerDtoToCsv.setDayOfTheWeek("Domingo");
+						plannerDtoToCsvList.add(plannerDtoToCsv);
+				  }
+				  if (planner.getMonday() != null) {
+						PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getMonday());
+						plannerDtoToCsv.setDayOfTheWeek("Lunes");
+						plannerDtoToCsvList.add(plannerDtoToCsv);
+				  }
+				  if (planner.getTuesday() != null) {
+						PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getTuesday());
+						plannerDtoToCsv.setDayOfTheWeek("Martes");
+						plannerDtoToCsvList.add(plannerDtoToCsv);
+				  }
+				  if (planner.getWednesday() != null) {
+						PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getWednesday());
+						plannerDtoToCsv.setDayOfTheWeek("Miércoles");
+						plannerDtoToCsvList.add(plannerDtoToCsv);
+				  }
+				  if (planner.getThursday() != null) {
+						PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getThursday());
+						plannerDtoToCsv.setDayOfTheWeek("Jueves");
+						plannerDtoToCsvList.add(plannerDtoToCsv);
+				  }
+				  if (planner.getFriday() != null) {
+						PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getFriday());
+						plannerDtoToCsv.setDayOfTheWeek("Viernes");
+						plannerDtoToCsvList.add(plannerDtoToCsv);
+				  }
+				  if (planner.getSaturday() != null) {
+						PlannerDtoToCsv plannerDtoToCsv = mapper.toPlannerDtoToCsv(planner.getSaturday());
+						plannerDtoToCsv.setDayOfTheWeek("Sábado");
+						plannerDtoToCsvList.add(plannerDtoToCsv);
+				  }
 			}
 
 			return plannerDtoToCsvList;
@@ -137,9 +147,36 @@ public class PlannerService {
 			plannerRepository.delete(planner);
 	  }
 
-	  public void deleteByUserId(String userId){
+	  public void deleteByUserId(String userId) {
 
 			Planner planner = plannerRepository.findOneByUser(userId);
-			plannerRepository.delete(planner);
+			if (planner != null) {
+				  plannerRepository.delete(planner);
+			}
+	  }
+
+	  private PlannerWithRecipeDTO mapperPlan(Planner planner) {
+
+			PlannerWithRecipeDTO plannerDTO = new PlannerWithRecipeDTO();
+			if (planner != null) {
+				  plannerDTO.setId(planner.getId());
+				  plannerDTO.setIdUser(planner.getIdUser());
+				  if (planner.getMonday() != null)
+						plannerDTO.setMonday(mapper.toFullRecipeDto(planner.getMonday()));
+				  if (planner.getTuesday() != null)
+						plannerDTO.setTuesday(mapper.toFullRecipeDto(planner.getTuesday()));
+				  if (planner.getWednesday() != null)
+						plannerDTO.setWednesday(mapper.toFullRecipeDto(planner.getWednesday()));
+				  if (planner.getThursday() != null)
+						plannerDTO.setThursday(mapper.toFullRecipeDto(planner.getThursday()));
+				  if (planner.getFriday() != null)
+						plannerDTO.setFriday(mapper.toFullRecipeDto(planner.getFriday()));
+				  if (planner.getSaturday() != null)
+						plannerDTO.setSaturday(mapper.toFullRecipeDto(planner.getSaturday()));
+				  if (planner.getSunday() != null)
+						plannerDTO.setSunday(mapper.toFullRecipeDto(planner.getSunday()));
+			}
+
+			return plannerDTO;
 	  }
 }
